feat: Add "Select Neovim" button to startup error notification
